+++ test_data/dyn-3.1.0/src/dyn/__init__.py
@@ -1,0 +1,5 @@
+__version__ = '3.1.0'
+
+def hello_dyn():
+    print("This is dyn 3.1.0")
+