+++ tests/test_payne.py
@@ -153,20 +153,24 @@ class TestPayneLocal:
             # assert process_output([bin_dir / "foo-1.3.2"]) == (
             #     "This is foo 1.3.2\nThis is bar 1.2.0\nThis is baz 1.1.0\n", "")
 
-    # Cannot install sup with locked dependencies because it has no lockfile
+    # Cannot install sup or dyn with locked dependencies because they have no
+    # lockfiles
     @pytest.mark.parametrize("locked", [False])
-    def test_install_local(self, locked):
+    @pytest.mark.parametrize("name, version", [
+        ("sup", "2.1.0"),
+        ("dyn", "3.1.0"),
+    ])
+    def test_install_local(self, locked, name, version):
         with TemporaryDirectory() as temp_dir:
             apps_dir = temp_dir / "apps"
             bin_dir = temp_dir / "bin"
 
             payne = Payne(apps_dir, bin_dir, {"payne_test_data": "http://localhost:8000/payne_test_data"})
 
-            # Install sup 2.1.0
-            payne.install_project(test_data / "sup-2.1.0", locked=locked, reinstall=False)
-            assert self.installed_apps(apps_dir) == {"sup": {"2.1.0"}}
-            assert self.installed_scripts(bin_dir) == {"sup-2.1.0"}
-            self.assert_app_valid(apps_dir, "sup", "2.1.0")
+            payne.install_project(test_data / f"{name}-{version}", locked=locked, reinstall=False)
+            assert self.installed_apps(apps_dir) == {name: {version}}
+            assert self.installed_scripts(bin_dir) == {f"{name}-{version}"}
+            self.assert_app_valid(apps_dir, name, version)
 
-            assert process_output([bin_dir / "sup-2.1.0"]) == (
-                "This is sup 2.1.0\n", "")
+            assert process_output([bin_dir / f"{name}-{version}"]) == (
+                f"This is {name} {version}\n", "")