+++ doc/changelog.md
@@ -2,6 +2,7 @@
   * Metadata file renamed from `payne_app.json` to `payne_app-version.json`
   * Script metadata extended
   * Schema version added to metadata
+  * Support for dynamic version in `pyproject.toml` added
 
 0.2.0:
   * Installing applications from PyPI